refactor: Extract PlanCards into a reusable component and add pricing section to LoginPage
